Move to signal module + function
