+++ crates/burn-tensor/src/tensor/signal/mod.rs
@@ -1,0 +1,3 @@
+mod hann_window;
+
+pub use hann_window::*;
